Updated file

+++ src/app.js
@@ -183,7 +183,7 @@ app.delete("/messages/:id", (req, res) => {
     .collection("messages")
     .findOne({ _id: new ObjectId(id) })
     .then((message) => {
-      //if (message.from !== user) return res.sendStatus(401);
+      if (message.from !== user) return res.sendStatus(401);
       database
         .collection("messages")
         .deleteOne({ _id: new ObjectId(id) }, { name: user })
